Kilogram added till BillHistoryActivity

+++ app/src/main/java/com/example/easy_billing/CartItem.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
 
 data class CartItem(
     val product: Product,
-    var quantity: Int
+    var quantity: Double
 ) : Serializable {
 
     fun subTotal(): Double {
